Initial Commit
This commit represents my first open-sourcing of the ArtContracts project!
This code state is being designated as Version 1.4, released on 2021/07/21

WARNING: The code in its current state does not include any sort of "README" or user documentation. In the coming days, I will make changes to thoroughly explain the intended use. The code itself is also subject to change during these edits, as I may make improvements to streamline the user experience.