Re-Enable Consecutive Token IDs
Thanks to @RodrigoQuelhas on Twitter/Telegram who showed me the specific template line which needs to be removed in order to allow proper "multiple minting" (of single Token IDs) when assume_consecutive_token_ids is equal to True.

NOTE: This is not yet manually tested, but if all my manual tests pass then it should be ready for mainnet deployment!